"Renamed index.jsx to app.jsx, renamed index.js, adjusted file paths"
